+++ pylops/basicoperators/SecondDerivative.py
@@ -39,7 +39,7 @@ class SecondDerivative(LinearOperator):
         y[i] = (x[i+1] - 2x[i] + x[i-1]) / dx
 
     """
-    def __init__(self, N, dims=None, dir=0, sampling=1, dtype='float32'):
+    def __init__(self, N, dims=None, dir=0, sampling=1, dtype='float64'):
         self.N = N
         self.dir = dir
         self.sampling = sampling
